pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lscal"
version = "1.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "icalendar>=6.3.2",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.5",
]

[tool.hatch.build.targets.wheel]
packages = ["src/lscal"]

[project.scripts]
lscal = "lscal.lscal:main"
